// apps/web/src/api-client/model/adminRideListItem.ts
/**
 * Generated by orval v7.21.0 🍺
 * Do not edit manually.
 * Waymate API
 * Carpooling backend powering the Waymate web app.
 * OpenAPI spec version: 0.1.0
 */
import type { RideId } from "./rideId";
import type { RideStatus } from "./rideStatus";
import type { Currency } from "./currency";
import type { AdminRideListItemDriver } from "./adminRideListItemDriver";

export interface AdminRideListItem {
    id: RideId;
    rideStatus: RideStatus;
    departureAt: string;
    /**
     * @minimum -9007199254740991
     * @maximum 9007199254740991
     */
    offeredSeats: number;
    currency: Currency;
    originCity: string;
    destinationCity: string;
    /**
     * @minimum -9007199254740991
     * @maximum 9007199254740991
     */
    activeSeatCount: number;
    driver: AdminRideListItemDriver;
    createdAt: string;
}